python_service/requirements.txt
# FastAPI и веб
fastapi
uvicorn[standard]
requests
python-dotenv

# Pydantic для валидации
pydantic

# AI/LLM
langchain
langchain-openai
langchain-core
openai